Next_stop/Agente_usuarios/requirements.txt
fastapi
uvicorn
requests
pydantic
google-generativeai
mirascope
pillow
flask